third_party/blink/web_tests/external/wpt/dom/traversal/TreeWalker-acceptNode-filter-cross-realm-null-browsing-context.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TreeWalker: NodeFilter from detached iframe works as expected</title>
<link rel="help" href="https://dom.spec.whatwg.org/#ref-for-call-a-user-objects-operation%E2%91%A0">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
const t = async_test();

const iframe = document.createElement("iframe");
iframe.src = "support/TreeWalker-acceptNode-filter-cross-realm-null-browsing-context-subframe.html";
iframe.onload = t.step_func_done(() => {
    const nodeIterator = iframe.contentWindow.createNodeIterator();
    iframe.remove();

    assert_equals(iframe.contentWindow, null);
    assert_equals(nodeIterator.nextNode(), document.body);
    assert_true(nodeIterator.dummyFilterCalled);
});

document.body.append(iframe);
</script>
